requirements.txt
openai==1.35.1
pdf2image==1.17.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.1